chore: chat 폴더 안으로 이동
