+++ src/app/ai-temp/page.tsx
@@ -1,4 +1,4 @@
-import AiChatPanel from '../components/ai/AiChatPanel';
+import AiChatPanel from '../components/ai/chat/AiChatPanel';
 
 export default function AITempPage() {
   return <AiChatPanel />;
