+++ src/app/components/ai/chat/AiChat.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import { useState } from 'react';
-import ActionBubble from './ActionBubble';
 import ChatBubble from './ChatBubble';
 import ChatInput from './ChatInput';
 
@@ -9,7 +8,6 @@ interface Message {
   id: number;
   role: 'user' | 'assistant';
   text: string;
-  actions?: string[];
 }
 
 export default function AIChat() {
@@ -18,7 +16,6 @@ export default function AIChat() {
       id: 1,
       role: 'assistant',
       text: 'that looks so good that looks so good that looks so good that looks so good that looks so good!',
-      actions: ['문제 교정', '내용 늘리기'],
     },
     {
       id: 2,
@@ -52,14 +49,9 @@ export default function AIChat() {
 
   return (
     <div className="flex flex-col h-full">
-      <div className="p-4 border-b bg-white text-lg font-semibold">텍톡 AI</div>
-
       <div className="flex-1 overflow-y-auto p-4 space-y-6">
         {messages.map((msg) => (
-          <div key={msg.id} className="flex flex-col gap-2">
-            <ChatBubble role={msg.role} text={msg.text} />
-            {msg.actions && <ActionBubble actions={msg.actions} />}
-          </div>
+          <ChatBubble key={msg.id} role={msg.role} text={msg.text} />
         ))}
       </div>
