+++ src/app/components/ai/chat/ChatBubble.tsx
