+++ src/app/components/ai/chat/ChatInput.tsx
